无

+++ src/practice/Prime.java
@@ -5,7 +5,7 @@ public class Prime {
 		int i, j, k;
 		int num = 0;
 		for (i = 1; i <= 100; i++) {//外层循环i控制1到100
-			k = (int) Math.sqrt((double) i);//求出i的平方根k
+			k = (int) Math.sqrt(i);//求出i的平方根k
 			for (j = 2; j <= k; j++) {//从2依次除到k，如果不能整除就是素数
 				if (i % j == 0) {
 					break;
